some change

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -1,10 +1,11 @@
 /**
  * _sqrt_recursion - this funcion calculates the square root of a number
  * @n: is the number which find the square root
+ * @i: is the second number
  * Return: the value
  */
 
-int _sqrt_helper(int n, int i);
+int _sqrt_helper(long int n, long int i);
 
 int _sqrt_recursion(int n)
 
@@ -23,7 +24,7 @@ int _sqrt_recursion(int n)
  * Return: the value
  */
 
-int _sqrt_helper(int n, int i)
+int _sqrt_helper(long int n, long int i)
 {
 	if (i * i > n)
 		return (_sqrt_helper(n, i - 1));
